BranchingProcesses.solve_and_split_constantrate — Function

solve_and_split_constantrate(prob::T, λ::S) where {T<:SDEProblem, S<:Real}

Recursively solve a branching stochastic process where the single-particle dynamics is defined by the SDE problem prob and the branching rate is a constant λ.

The timespan of the problem prob defines the total time interval for the branching process. A lifetime for the first particle is sampled from an exponential distribution with rate λ. If the lifetime is larger than the total time interval, the problem is solved until the end of the interval and a solution node is returned without children. If the lifetime is smaller than the total time interval, the problem is solved until the sampled lifetime, and a solution node is returned with recursively solved children for the remaining time interval.

See also sample_lifetime_constantrate, SDE problems.

BranchingProcesses.sample_lifetime_constantrate — Function

sample_lifetime_constantrate(λ::T) where T <: Real

Sample the lifetime of a particle when the branching rate is a constant λ independent of time or the value of the process. This is equivalent to sampling from an exponential distribution with rate λ.

Note that this is not the same as a ConstantRateJump where the rate is only constant between jumps.