Types

Problem types

BranchingProcesses.BranchingSDEProblem — Type

struct BranchingSDEProblem{P<:SciMLBase.SDEProblem, R<:Union{Function, Real}, O<:Union{Integer, Distributions.DiscreteUnivariateDistribution, Distributions.UnivariateDistribution{Distributions.Discrete}, Distributions.DiscreteDistribution{Distributions.Univariate}}} <: BranchingProcesses.BranchingProblem

A structure to define a branching stochastic process.

Fields

prob: The SDE problem defining the single-particle dynamics of the branching process, an instance of SciMLBase.SDEProblem.

branchrate: The branching rate of the process, which can be a constant or a function of time and/or the state of the process. If a constant, it must be a positive real number.

nchild: The number of children to be created for each particle, which can be a non-negative integer or a discrete distribution with non-negative support from which the number of children is sampled.

Solution types

BranchingProcesses.BranchingProcessSolution — Type

struct BranchingProcessSolution{T<:Union{SciMLBase.DESolution, SciMLBase.SciMLSolution}}

A tree structure to hold a solution of a branching stochastic process.

Fields

sol: The solution, that is, sample trajectory, of the particle associated to the node, which is an instance of SciMLBase.AbstractSciMLSolution.

children: a vector of BranchingProcessSolution instances representing the child particles of the current particle.

The solution type T must be the same for all nodes in the tree.

AbstractTrees.children — Function

AbstractTrees.children(node::BranchingProcessSolution)

Return the children of a BranchingProcessSolution node.

AbstractTrees.nodevalue — Function

AbstractTrees.nodevalue(node::BranchingProcessSolution)

Return the value of a BranchingProcessSolution node, defined as its final state value.